requirements.txt
streamlit==1.42.2
python-dotenv==1.0.1
psycopg2-binary
langchain==0.3.19
langgraph==0.2.74
langchain-openai==0.3.7
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
pydantic